GittinsIndices

Documentation for GittinsIndices.

GittinsIndices.calculate_bernoulli_gittins

GittinsIndices.calculate_discrete_gittins

GittinsIndices.calculate_gaussian_gittins

GittinsIndices.calculate_bernoulli_gittins — Method

Calculates bernoulli gittins indices for a given arm prior, parameterized as Beta(alpha, beta). To calculate all the possible states from a given starting prior and a number of pulls, optionally pass the pulls argument. the function will then return a triangular matrix where gi_matrix[i, j] = gittins index of beta(alpha + i - 1, beta + j - 1) for i, j >= 1 and i + j <= pulls.

alpha: current successes
beta: current failures
gamma: discount factor
pulls: optional arg that tells the function how many priors to calculate

Optional args controlling accuracy of approximation: N: how far to hallucinate into the future tol: fineness of approximation given all of the above.

GittinsIndices.calculate_discrete_gittins — Method

Given the specified discrete stochastic process, compute the gittins index of the starting state.

m: number of states in stochastic process P: transition matrix between states r: reward vector with each entry specifying reward for that state initial_state: starting state, whose gittins index will be returned gamma: discount factor

GittinsIndices.calculate_gaussian_gittins — Method

Calculate the gittins index for a given gaussian belief. Takes the following required parameters: mu: mean of gaussian tau: precision (inverse of variance) n: number of times the arm has been pulled gamma: discount factor. Also optionally parameters concerning the accuracy of the returned approximation: xi: controls the range of the discretization, i.e. many standard deviations away from the mean to consider discretizing delta: the step size in the discretization of the gaussian N: horizon size, i.e. how many states you hallucinate into the future tol: the fineness of the calculation, given the above params